refactor: make `semantic` internal
