+++ generators/generate-debian-versions.py
@@ -179,7 +179,7 @@ def fetch_packages_versions():
   return extract_packages_with_versions(osvs)
 
 
-outfile = "pkg/semantic/fixtures/debian-versions-generated.txt"
+outfile = "internal/semantic/fixtures/debian-versions-generated.txt"
 
 packs = fetch_packages_versions()
 with open(outfile, "w") as f:
