+++ generators/generate-pypi-versions.py
@@ -101,7 +101,7 @@ def fetch_packages_versions():
   return extract_packages_with_versions(osvs)
 
 
-outfile = "pkg/semantic/fixtures/pypi-versions-generated.txt"
+outfile = "internal/semantic/fixtures/pypi-versions-generated.txt"
 
 packs = fetch_packages_versions()
 with open(outfile, "w") as f:
