+++ internal/semantic/fixtures/maven-versions.txt
