+++ internal/semantic/fixtures/nuget-versions.txt
