+++ internal/semantic/utilities.go
